Fix for OK button issue
OK button was not working due to an error in the settings that were being applied. The setting for listenign to a different parameter for updates would be undefined if there weren't any other parameters available to populate the list. This caused an issue when trying to set the setting to undefined.