Fixed filter option
Option to update the list based on mark selection or filter change was not tied in and useless. Updated it to actually function based on the options set. If you currently have the extension in use it will work as normal and will set the option to default to "on".